Split 'messages.js' to several files
It's make maintenance easier.
