+++ inspector/static/inspector/js/messages/message_details.hbs
