+++ inspector/static/inspector/js/messages/models.js
@@ -1,8 +1,6 @@
 import Backbone from 'backbone';
 import {_} from 'underscore';
 
-import messageDetailsTemplate from 'inspector/js/message_details.hbs!';
-
 
 export class Message extends Backbone.Model {
 
@@ -67,37 +65,5 @@ export class Message extends Backbone.Model {
   get sound() {
     return soundManager.sounds[this.soundId];
   }
-
-}
-
-export class MessageDetailsView extends Backbone.View {
-
-  get events() {
-    return {
-      'click .play-button': 'playSound'
-    };
-  }
-
-  initialize() {
-    this.listenTo(this.model, 'change', this.render);
-    this.listenTo(this.model, 'destroy', this.remove);
-  }
-
-  render() {
-    let html = messageDetailsTemplate({'message': this.model});
-    this.$el.html(html);
-    return this;
-  }
-
-  remove() {
-    this.undelegateEvents();
-    this.stopListening();
-    //this.$el.empty();
-  }
-
-  playSound() {
-    this.model.sound.play();
-  }
-
 
 }
